Use mutable C arrays except for strings
These changes introduce a mutable representation of a C array and
retains the existing support for the immutable representation.  The
mutable representation is used for all C arrays that are not strings.

The existing implementations of functors CArray and CArrayN for MLton
allow a C array to be stored as an SML vector because MLton allows an
SML vector to be passed to a C function where it expects a C array,
provided that the C function does not modify a C array argument because
MLton requires that the SML vector is not modified.  Therefore, these
functors create a C interface for an immutable C array.

An immutable representation of C arrays has worked well so far because
there do not appear to be many, if any, cases of a C function that
modifies a C array parameter whose direction is 'in' or 'inout' and
whose ownership is not transferred.  (For a C array parameter not
satisfying these conditions, the bindings would internally provide a
new or duplicate C array to the C function so the original is not
modified.)  Notably, these conditions would not be satisfied for an
instance parameter but a C array can never be an instance parameter.
However, it is possible for a library to provide a C function with
introspection annotations that result in a C array being destructively
updated.  Therefore, mutable C arrays are generally required.

The existing functors CArray[N], which construct a C interface for an
immutable C array are renamed to ConstCArray[N] and new functors
CArray[N] are introduced to construct a C interface for a mutable
C array:

    CArray               --->  ConstCArray
    CArrayN              --->  ConstCArrayN
    <>                   --->  CArray
    <>                   --->  CArrayN

In the Poly/ML implementation, the code common to ConstCArray[N] and
CArray[N] is factored out into an internal functor CArray[N]Common.

There is scope for generating bindings that use the immutable
representation of a C array where possible because the GIR files include
the C type of an array and the const declarator indicates whether
elements are not modified.  However, the current abstract SML types
for a C array do not allow a non-const C array to be used where a const
C array is expected.  Therefore the representation of a C array is fixed
for a particular type and all types are mutable.  (The types 'utf8' and
'filename', which are distinct from C arrays in the introspection types,
are immutable, necessarily so because their abstract representation is
the SML string type.)  Using an immutable representation for a const
C array could be done in future because this would not break the API.

No change is required to codegen.