Fix omission in codegen design and C interface for a 'value' record
In the function codegen design document, withStructFun and fromStructFun
are fixed in the case

  direction = INOUT
  isPointer = false
  ownershipTransfer = EVERYTHING

to respect optionality.  The signature VALUE_RECORD is updated to
provide the high-level FFI function `withDupOptPtr`.  No change is
required to the codegen code which would already output `withDupOptPtr`
as required, but there are no examples of this case currently.

Note that the signature RECORD, which provides the C interface for a
'non-value' record, does not need the high-level FFI functions used
only when 'isPointer' is false because 'isPointer' is always true for a
'non-value' record.