Refactor C interface for callback types
The signature CALLBACK is updated to represent the C interface for all
callback types.  The signature NOTIFIED_CALLBACK is removed.  (Note that
the scope of a callback is not fixed for each callback type.)

The functor Callback is introduced and used to construct the C interface
of the currently supported callback types (GLib.SourceFunc, GLib.IOFunc,
GLib.ChildWatchFunc and GLib.SpawnChildSetupFunc).  The C interface
supports all scope values by updating some high-level FFI functions with
an extra parameter that capture the scope.  As a result, the closure for
a callback with scope 'call' and 'async' is now freed as required, and
no longer leaked.

The signature CLOSURE is introduced to encapsulate a closure and the
functor Callback takes an implementation of this as an argument.  The
functor Closure is introduced to construct implementations of CLOSURE.
In addition to implementing callback types, the implementation of signal
handlers is refactored to use this, which given better alignment of the
MLton and Poly/ML code.

In the MLton version of the functor Closure, the exception history is
printed if the enclosed SML function raises an exception.  This has no
effect if the MLton compiler is not used with the option

    -const 'Exn.keepHistory true'

Poly/ML does not appear to have an support for printing the exception
history.  (The functions for tracing an exception in PolyML.Exception do
not appear to do anything.)