Remove non-standard interface to GLib.childWatchSourceNew
GLib.childWatchSourceNew is not required because GLib.childWatchAdd
provides the required behaviour so it is reverted to its generated
version.  (Note that the generated version for this source and other
sources, GLib.<X>SourceNew, are useless because GLib.Source.setCallback
is not available, and cannot be provided in a type-safe manner unless
GLib.Source.t is parameterized by the callback function type.)  The
C function that provides the implementation of GLib.childWatchSourceNew
is also removed, along with variants for other sources that were not
used.

The manually generated bindings for GLib.childWatchAdd, GLib.timeoutAdd
and GLib.timeoutAddSeconds are fixed to call the 'full' variant of the
C function internally.