Refactor signature C_POINTER using C_MEMORY and other minor changes
The operations in C_POINTER that perform memory allocation and pointer
arithmetic are replaced by a structure Memory that implements C_MEMORY.

Also, in C_POINTER, the structure OptValueType is changed to implement
C_VALUE_EQ_NULL_TYPE where type `p` is a non-optional pointer.  Also,
the structure NonOptValueType is renamed to ValueType for consistency.