Use MLton callback mechanism with Poly/ML
With this approach, the closure is an index into a callback table.  A
Poly/ML closure is defined once for each dispatch and notify function of
each callback type and a GObject closure.  The lifespan of these Poly/ML
closures is the life of the application, so they never need to be freed
(and so it does not matter that there is no mechanism to free a Poly/ML
closure).