Fix ownership transfer for non-pointer 'inout' parameters
Prior to this change, an 'inout' parameter that is a non-pointer, i.e.
has only one level of indirection in its C type, was:

 1. not supported in the case of transfer mode 'none';
 2. incorrectly supported for arrays in the case of transfer mode
    'container' because ownership of the array elements was also
    passed to the C function and the memory allocated for the
    elements subsequently leaked.

Note that the current bindings are not affected by these issues because
only ownership transfer mode 'full' is used for such a parameter.

All cases of ownership transfer are supported for a non-pointer 'inout'
parameter by updating the high-level FFI so that `withDup[Opt]Ptr` has a
transfer argument as for `with[Opt]Ptr`.