Deprecate high-level FFI modules ConstCArray[N]
The purpose of these modules is to enable a C function argument to
directly reference an array in the ML heap, to avoid making a temporary
copy on the C heap, when using MLton.  It has been observed that this is
not compatible with managing ownership of memory from SML because an
address in the ML heap is not stable - it may change at any time due to
a GC that relocates data - so it is not possible to copy from an address
in the ML heap in SML code.  Although not all functions are affected,
there is an issue in the following cases:

 1. A C function returns a pointer into the ML heap, which can happen
    when the C function is given a pointer into the ML heap.  By the
    time `FFI.[from|copy][Opt]ptr d` is applied to copy data to the
    C heap, a GC could have relocated the data in the ML heap.

 2. `FFI.with[Ref|Dup][Opt]Ptr d` is used when `d > 0`.  This requires
    the initial `d` levels to be copied from the ML heap but level `d`
    would reference addresses in the ML heap for levels below `d` and
    those addresses may become invalid due to a GC that relocates the
    data.  In practice, this case occurs only when `d` is 1, which
    corresponds to the ownership transfer mode 'container'.

 3. `FFI.withDup[Opt]Ptr d` is used when `d = 0`.  This requires a
    pointer to the vector in the ML heap to copy after the function has
    returned but the pointer may become invalid due to a GC that
    relocates the data.  In practice, this case occurs only for a non-
    pointer 'inout' array parameter where the ownership transfer mode is
    'none'.  (This an unusual case anyway and there are no known
    examples.)

Cases 2 and 3 above are detected and will result in an exception at
compile time.  Case 1 cannot be detected.

Prior to this change, ConstCArray was used only to implement Utf8.  Its
use there is replaced by CArray.