Prevent finalization of function arguments until call completes
A C function may pass back a pointer to data whose ownership is not
transferred, requiring the data to be duplicated (either copied or a
reference added).  It is possible that the data to be duplicated is
owned by one of the arguments of the C function and, furthermore, the
last reference to the data is in the argument of the C function.
Consequently, if a GC occurs after the C function returns but before the
data has been duplicated, it is possible that the data is destroyed
before it has been duplicated.

This issue is unlikely to occur in practice but is theoretically
possible.  The issue is fixed by touching every instance, 'in' and
'inout' pointer argument of the C function after all 'from' functions
have been evaluated, if there are any 'from' functions duplicate any
data.  This may touch some arguments unnecessarily but the worst case
must be assumed as nothing is known about data dependencies within the
C function.

This change also fixes codegen for the high-level function <withSelfFun>
which always produced `withPtr` and now generates `withVal` for a method
of an enum or flags interface.