Add struct type for union field and partial codegen for union interface
The union interface codegen design is not finalised and the current
approach used with Gdk.Event may have some issues.  With this change,
a record that is used solely as the type of a field in a union interface
can be designated as a union record in the configuration and code is
generated based on the current approach used with Gdk.Event, providing
much of the existing manual code.

An obvious limitation of the current design is that a record used for
the type of a union field is assumed to be used nowhere else.