Remove generated code
There is no need to keep the generated code under version control
because the versions of GIR and TYPELIB files used are configured, which
should allow the generated code to be reliably created.

The codegen script is modified to create the directories auto-0-gir and
auto-0-typelib if they do not already exist.  Similarly, the script
mlton-codegen is modified to create the directory mlton-auto if it does
not already exist.