Expose array C pointer type implementation as C.PointerType
Prior to this change, the C interface for an array provided the sub-
structure C.ArrayType which implements the signature C_ARRAY_TYPE and,
therefore, C_POINTER_TYPE when `type 'a from_p = 'a`.  Consequently, to
refer to the C pointer type implementation of a C interface structure,
it was necessary to refer to the sub-structure C.ArrayType for a C array
and C.PointerType otherwise.  This change allows the C pointer type
implementation of any C interface to be referred to by the sub-structure
C.PointerType (as well as C.ArrayType, which is still required by the
functors CArray[N] to construct a C array FFI).

This change generalizes the signature C_POINTER_TYPE by introducing the
type `'a from_p` as used in the signature C_ARRAY_TYPE.  Although
C_ARRAY_TYPE now includes C_POINTER_TYPE, this cannot be expressed via
an include specification because the MLton variant of C_POINTER_TYPE
contains a sub-structure Vector which would need to be extended in the
MLton variant of C_ARRAY_TYPE, which is not possible.

This allows a small simplification to the code generator which no longer
needs to check for an array when generating a reference to a C pointer
type structure because C.PointerType is referenced in all cases.