Extend C interface to guarantee conversion to/from C pointer
In a C interface for a pointer type, it is not necessarily the case that
type `t` is the same as type `C.PointerType.t` because a C interface may
require type `t` to store extra information or to allow type `t` to use
a different representation.  For both these reasons, this is the case
for the array C interface provided by the signature C_ARRAY.  In this
case, the C sub-structure of a C interface provides no means to convert
between the type `t` and a C pointer.  (When these types are the same,
`C.PointerType.toC` and `C.PointerType.fromC` can be used.)  For many C
interfaces, functions in the FFI sub-structure provide such conversions
but only if the C pointer type `... C.p` is used in the low-level FFI.
This is not the case for the array C interface, which uses the type
`... FFI.in_p` for function arguments (which are transformed by a C-side
wrapper function).  Consequently, the array C interface provides no way
to convert between type `t` and a C pointer.  Such a conversion is
required to get and set an array field of a record or class.

The C sub-structure of every C interface specification is extended with
the following functions:

    val takePtr : non_opt p -> t from_p
    val giveDupPtr : (non_opt p -> 'a) from_p -> t -> 'a
    val touchPtr : t -> unit

If the implementation of type `t` uses the C representation, i.e. has
a pointer to data that it owns in the C heap, then the following
function is also included:

    val withPtr : (non_opt p -> 'a) from p -> t -> 'a