Add support for fields
'get' and 'set' accessor functions are generated for each field of a
record that is neither private nor a bit field.  'get' is provided only
if the field is readable and 'set' is provided only if the field is
writable (and not the length of an array field).  A field of a record
may be excluded by configuration.

Although a class may have fields, accessor functions are generated for
a field of a class only if included by configuration.