Fix C interface for pointer values to avoid loss of pointer type safety
In the signature C_POINTER, sub-structures ValueType and OptValueType
provide the C interface for pointer values.  For each of these
sub-structures, the type `p` was defined to be the same non-optional
pointer as the type `v`.  Consequently, type checking considered these
types to be equal and would not catch certain incorrect use of pointers.
The only incorrect use in BoxedValueRecord was inconsequential because
it was hidden behind the VALUE_RECORD signature.  This issue is resolved
by leaving type `p` abstract for each sub-structure.