Add configuration to specify that a pointer field is non-optional
In the absence of GObject Introspection annotations to indicate whether
a field is optional, a field that is a C pointer is assumed to be
optional.  This new configuration parameter allows such fields to be
treated as non-optional.

The configuration is updated to specify that GLib.Error.message is a
non-optional field.