Use nullable annotations for 'out' and 'inout' parameters
These result of `GIRepository.ArgInfo.mayBeNull` was ignored for 'out'
and 'inout' parameters because the value used not to be meaningful.
Now that the nullable annotation is well established and supported by
GObject Introspection, this value is now considered meaningful.

Also, the condition under which an array has an optional type is updated
to ensure that an array that can be empty, i.e requires no bytes, is not
optional (because the empty array may be represented by a null pointer).