Fix build failure due to GLib deprecated macros
The GLib macros
  - GLIB_DEPRECATED_MACRO
  - GLIB_DEPRECATED_MACRO_FOR
cause a GCC warning that is not covered by -Wno-deprecated-declarations
and the build fails due to -Werror.  To work around this, CFLAGS passed
to all uses of GCC includes -DGLIB_DISABLE_DEPRECATION_WARNINGS which
disables all GLib deprecation macros, including those mentioned above.

The failure occurs since use of these macros was introduced in 2.62.0.

Note that -Wno-deprecated-declarations is still required for other
libraries.