Rename signature UTF8 to UTF_8
UTF8 does not fit the UCU name format used for signature names.  This
breaks the code that generates the signature for an alias type defined
to be a UTF-8 string.  There is no example of this currently but there
is in a later version of GLib, for example the alias type GLib.Strv.
The generated code is not affected.