Check alias interface type is supported
An alias type that has an unsupported interface type, e.g. a callback,
should be excluded.  Without this check, an unsupported alias interface
type would not be excluded even though its (non-alias) source type would
be excluded.  Consequently, the source files for the alias type would
have a missing dependency reported.  (Although this results in the
source files being excluded, this is not recorded in the excluded log
and we should not see this error message.)

There are no examples of this currently so the generated code and the
excluded log are not affected.