Fix codegen for value record with no fields
Although a record with no fields cannot be automatically classified as a
value record, it is possible to manually configure such record to be a
value record.

For a value record with no fields, the function `newStrDecHighLevel`
raised the exception Empty.  The codegen is fixed in this case.  The
generated code is adjusted to have an equivalent form.