Use configuration to exclude functions in MLton codegen
This avoids generating C wrappers for functions that are not required.

Note that the implementation of GLibQuark.fromString requires the
generated wrapper 'mlton_g_quark_from_string'.  This would no longer be
generated because the function symbol 'g_quark_from_string' is excluded.
There is no need to exclude the functions 'g_quark_from_string' and
'g_quark_to_string' so these are no longer excluded.