Install an empty MLB file for each unavailable library
The installation process writes an empty 'mlton.mlb' file for each
library that is unavailable.  This ensures that excluding the SML files
for an unavailable library does not break MLB file dependencies of
libraries that are still included.

This is useful in the case that a library cannot be found (because,
for example, its development package is not installed) and it is not
wanted.  Also, this could be needed if the introspection dependencies
(from a GIR file) include more namespaces than the libraries actually
require when installed on a system with older libraries.