Fix conversion between MLton.Pointer.t and SysWord.word (for MLton)
A pointer `p` was converted to a word using

  MLton.Pointer.diff (p, MLton.Pointer.null)

This breaks when `MLton.Pointer.sizeOfPointer * 8 > Word.wordSize`.  In
practice, this occurs on x86_64 platforms because pointers are 64 bits
but `Word.wordSize` is (by default) 32.  This is fixed by using

  C_Pointer.diff (p, MLton.Pointer.null)

instead because the result has type `C_Ptrdiff.int` whose range is wide
enough for the result.