Extend signature C_MEMORY with conversion to and from CMemory.Pointer.t
The signature C_MEMORY is extended with functions `Pointer.toPointer`
and `Pointer.fromPointer` that convert the local type `Pointer.t` (which
may be abstract) to and from, respectively, the type `CMemory.Pointer.t`
which represents an arbitrary pointer.

Previously, conversion between abstract pointers of different types was
possible by converting via `SysWord.word`.  This change allows arbitrary
conversion without changing the internal representation, so this is
guaranteed to have no overhead.  This is used in the code that adds a
field offset to a struct pointer.