Add variant constructors to text trees
This allows text trees to embed variant content where each variant is
identified by a version, allowing different output to be generated for
different versions.

The signature VARIANT specifies a variant type and supporting functions
and is implemented by:
  - the functor ListVariant, given an implementation of VERSION;
  - the structure NoVariant.
In the structure NoVariant, the type `version` is uninhabited so no
variant version can be specified, thus there is only a default value.
The structure StringVersion provides an implementation of VERSION where
the type `version` is `string`.

The structure Variant provides the implementation of VERSION used in the
text trees.  In this change, Variant is defined as NoVariant, which does
not allow variants to be used, so the generated code is not affected.

The variant type is used to define a new case in each text tree data
type.  Text tree functions are updated to work with the variant cases
and new functions added.

The code generator is updated to produce output files for each possible
variant, using the variant version or 'default' as a suffix unless there
are no variant versions in which case there is no suffix.  The release
Makefile is updated to install the correct variant of MLton source files
(both MLB and SML) based on the MLton version detected by the configure
script.