Use a variant for each MLton import attribute
The signature CONCRETE_SYNTAX_TREE is modified to make a MLton import
attribute a variant optional value.  For each attribute, the function
`mkMLtonImportExp` introduces a variant that removes the attribute for
MLton versions where it is not available.

No import attributes are currently specified so the generated code is
not affected.