Hide MLton warning about constructor redefined by fun
This warning occurs in MLton releases after 20130715 due to the
following code that is used to bind the expression exp to the value
id when id is the name of a constructor:

    local
      fun <id> () = <exp>
    in
      val <id> = <id> ()
    end

In most cases, if not all, this occurs because there is a flags
value called 'NONE'.

An alternative approach is to use an 'open' declaration to remove the
constructor status of an identifier without causing a warning from MLton
as follows:

    local
      structure S : sig type t val <id> : t end =
        struct datatype t = <id> end
      open S
    in
      val <id> = <id> ()
    end

This approach is not used because it may result in reports about unused
code because the value defined in structure S is not referenced.

Therefore, the MLton check is disabled for the SML structure files in
each namespace by specifying the annotation 'valrecConstr ignore' in MLB
files for MLton versions greater than 20130715.