Use source files in release Makefile only for requested compilers
Prior to this change, the release Makefile created symbolic links for
all variant source files and installed all source files, even if the
source file was specific to a compiler not included by the configure
script.  This change ensures that the release Makefile builds and
installs source files only for the compilers included by the configure
script.

Most compiler-specific source files are in a compiler-specific directory
in a namespace directory but 'basis' source files (e.g. mlton.mlb,
polyml.sml) are not.  In order to distinguish these from portable source
files easily, the naming convention on 'basis' source files is tightened
to require their prefix to be the compiler identifier.  This requires
the following files to be renamed in the 'release' directory:

    src/gir/types-mlton.mlb   --->  src/gir/mlton-types.mlb
    src/gir/types-polyml.sml  --->  src/gir/polyml-types.sml