Check for new finalizers introduced by finalizers run on exit
When a finalizer is run, it can introduce a new finalizable value which
will add a finalizer to the global pending list.  This works as expected
during normal operation because `threadFn` in the structure Finalizable
runs the finalizers outside the critical section in `updatePendingList`
using:

  app run (updatePendingList clean ())

On exit, the function `doGCAndFinalize` is used to run all remaining
finalizers.  If these finalizers create any finalizable values, the new
finalizers added to the global pending list are discarded without being
run.  This change fixes the function `doGCAndFinalize` to check for new
finalizers on each iteration.