Support variant Poly/ML source files
The release Makefile is updated to install the correct variant of
Poly/ML source files based on the Poly/ML version detected by the
configure script.

The suffix of a variant Poly/ML source file is a hierarchical version
using '-' to separate levels.  For example:
  - file.sml.5-8-1   # used up to 5.8.1, not used for 5.8.2, 5.9.0, ...
  - file.sml.5-8-x   # used up to 5.8.x, not used for 5.9.0, ...
  - file.sml.default # used otherwise
Note that 'x' in '5-8-x' can be any letter because all letters are
considered greater than numbers.

In this change, no variant Poly/ML source files are introduced.