Add support for dynamic linking if available in Poly/ML
For versions of Poly/ML >= 5.8, dynamic linking is used for code
generated from GIR files.  This is achieved by replacing references to
`getSymbol` (as defined in the structure PolyMLFFI) by either
  - externalFunctionSymbol or
  - externalDataSymbol
(as also defined in the structure PolyMLFFI).  For versions of Poly/ML <
5.8, these functions are defined as `getSymbol` to fall back on dynamic
loading.  Dynamic loading is still used for code generated from TYPELIB
files.

The application of `externalFunctionSymbol` and `externalDataSymbol`
must be evaluated at compile time.  Therefore, code generation of the
structure for a record type is modified to ensure this.