Fix property type safety
The type `('object_class, 'get, 'set, 'init) Property.t` is not abstract
and the phantom type parameter `'object_class` was not mentioned in the
concrete definition.  Consequently, type checking did not ensure that
access to a property was allowed only for an object with the property.

In the signature PROPERTY, type `('object_class, 'get, 'set, 'init) t`
is fixed by explicitly referencing `'object_class' in the concrete
representation.  Further explanation is given in the new comments for
the type.