Use class and union base types in high-level FFI functions
This is done to allow high-level FFI expressions to be evaluated at
compile time.

The high-level FFI functions `with[Ref][Opt]Ptr` for a class and union
contain a variable in the type of the parameter: for a class, the type
variable in `'a class` allows an object of any subclass of the class;
for a union, the type variable in `'a union` allows a union instance to
have the type of any field of the union.  The presence of these type
variables means that an expression defining a high-level FFI cannot be
evaluated at compile time due to the value restriction.  To work around
the value restriction, `with[Ref][Opt]Ptr` functions are changed to use
the type `base class` for the class parameter and `base union` for the
union parameter.  In order to implement the existing specifications, a
class or union parameter is explicitly converted using the respective
`toBase` function in the code generated for a function.  Although this
is done at run time, there is no overhead because a `toBase` function is
just the identity function internally.  Finally, in the code generated
for a function, the high-level FFI expression is defined in a local
declaration outside the function so that it is evaluated at compile
time.