Use class and union base types in value accessors
This is done to allow a closure marshal expression to be evaluated at
compile time.  However, the generated code is not updated to evaluate
closure marshal expressions at compile time because this will be done by
a subsequent commit that changes the representation of signals.

The value accessors for a class and union contain a type variable in the
accessor write type: for a class, the type variable in `'a class` allows
an object value to be set to an object of any subclass of the class; for
a union, the type variable in `'a union` allows a boxed value to be set
to an instance of any field of the union.  The presence of these type
variables means that an expression defining a closure marshaller cannot
be evaluated at compile time due to the value restriction.  To work
around the value restriction, value accessors are changed to use the
type `base class` for the class write type and `base union` for the
union write type.  As a result, the write type of a value accessor is
always the same as the read type, so a value accessor no longer needs
two type parameters for the read and write types.  The representation of
a value accessor is updated to have a single type parameter.  In order
to implement existing specifications, a class or union instance is
explicitly converted in the generated code using the respective `toBase`
function where it occurs as an 'out' parameter or return value of a
signal and as an argument of a property 'set' or 'init' function.