Use new callback type in ClosureMarshal high-level FFI functions
This is done to allow high-level FFI expressions to be evaluated at
compile time.

The high-level FFI functions `ClosureMarshal.with[Opt]Ptr` contain a
variable in the type of the closure parameter, given as a closure
marshaller of type `('a -> 'b) marshaller` and a handler function of
type `'a -> 'b`.  The presence of these type variables means that an
expression defining a high-level FFI using `ClosureMarshal.with[Opt]Ptr`
cannot be evaluated at compile time due to the value restriction.  To
work around the value restriction, `ClosureMarshal.with[Opt]Ptr` is
changed to take a new type `ClosureMarshal.callback`, which is
constructed from a marshaller and handler by a new polymorphic function
`ClosureMarshal.makeCallback`.