Expose C pointer element representation in C array interface
In the signature C_ARRAY, the sub-structure C.Pointer is changed to
be an instance of C_TYPED_POINTER instead of just C_POINTER.  Similarly
for the sub-structure Pointer in the signature C_ARRAY_TYPE.

The functors that construct implementations of C_ARRAY are updated to
specify the type `C.Pointer.e`, which is not necessarily the same as the
type `C.e`.  Similarly, the functors that construct implementations of
C_ARRAY_TYPE are updated to specify the type `Pointer.e` which is not
necessarily the same as the type `e`.