Refactor code that generates SML specifications
The code that generates SML specifications for functions and signals is
refactored to defer the introduction of type variables to allow both
read and write types to be generated for signal arguments and for a
signal return value.  (This is needed to support signal emission.)

The code that generates SML specifications for properties and fields is
also updated to improve code reuse and for consistency.

This change fixes a potential issue in the generated code.  For a class
or union 'inout' parameter of a function, the result type used the same
type variable as the argument type which would allow the caller to
incorrectly assume that a returned object or union instance is of the
same subclass or field, respectively, as the object or union instance in
the argument for the same parameter.  However, the generated code is not
affected because, in the current GIR library metadata, there are no
class nor union 'inout' parameters.