Add support for calling a closure and for signal emission
The representation of a closure marshaller is updated to enable a
closure to be called.  Currently there is support only for handling the
call of a closure which reads the argument from the 'in' parameters and
writes the result to the 'out' parameters and return value of a closure.
This change adds support for writing the argument to the 'in' parameters
and reading the result from the 'out' parameters and return value of a
closure.  The functions to construct a marshaller are updated to enable
a single description of a marshaller to provide both.  The type of a
marshaller is updated to have new type parameters for the argument that
is written to a closure and result that is read from a closure because
these differ, respectively, from the argument type that is read from a
closure and result type that is written to a closure where they contain
class or union types.

The function `Signal.emit` is introduced to support signal emission.
As done for a closure marshaller, the type of a signal is updated with
new type parameters for the argument and result types for emission,
which in general may differ, respectively, from the argument and result
types for handling where they contain class or union types.