Update internal representation of C size types
There are two issues with the current representation of C size types:

 1. The signature C_MEMORY specifies that the size argument of the
    allocation functions is the SML type `word`.  This may not be the
    most appropriate SML type to represent a C value of type size_t for
    a particular compiler leading to unnecessary conversion via the type
    `word`.

 2. It is assumed that the size of C size types (`size_t` and `ssize_t`)
    are the same as C long integer types.  This is true on systems
    using ILP32 and LP64 data models (most Unix-like systems) but not
    those using LLP64 (Windows 64 bit).  Therefore the library is
    not portable to LLP64 systems.

To address issue 1, a compiler-specific type is allowed for the size
argument of `malloc[0]` in the signature C_MEMORY.  This type is
constrained to be the same as the internal representation (type `v`)
of the C value type constructed by the functor SizeType.  This avoids
unnecessary conversion via intermediate types, particularly in the
functor BoxedValueRecord in the declaration of `new0_`.

To address issue 2, the functors SizeType and SSizeType no longer use
the representation of C long integer types for size types.  Under MLton,
`C_Size.t` is used in both SizeType and SSizeType.  Note that there is
no structure `C_SSize` corresponding to the POSIX type `ssize_t`.  This
does not matter because conversion to the external `int` representation
sign extends.  Under Poly/ML, the structure PolyMLFFI is extended with
the conversions `cSize[Large]` and `cSsize[Large]` and these are used
in SizeType and SSizeType, respectively.  The structure PolyMLFFI
assumes the size types have same size as a pointer types in order to
define the conversions `cS[s]ize[Large]` in terms of `cPointer`.  This
assumption is consistent with ILP32, LP64 and LLP64 data models so is
valid for most Unix-like systems and Windows.  This assumption can only
be removed only once Poly/ML provides conversions for size types.