Add function to convert the object or interface type of a signal
The function `Signal.conv` is added to convert a signal on an object
or interface of type `'b` to a signal on an object or interface of
type `'a`, given a function of type `'a -> 'b`.

This allows a signal of an interface to be converted to a signal of
an object that implements the interface because both an object and
interface provide a function `as<Iface>` for each interface Iface
that they implement.