Update repository README file
A banner and overview have been added to clarify what the repository
provides and to explain that users do not need the repository because
the release packages and documentation are available on the website.
This should help avoid the confusion if a user starts from the Github
page as in https://github.com/giraffelibrary/giraffe/issues/2 .

The list of repository contents was missing some files which are added.