Remove ValueAccessor module from GObject namespace
The ValueAccessor module was originally in the GObject namespace but was
moved to the initial GLib namespace.  A structure and signature for the
ValueAccessor module was retained in the GObject namespace that added
nothing to those in the initial GLib namespace, so they can be removed.
(Note that signatures in the GLib namespace are already referring to the
global type `ValueAccessor.t`, so there is no point in signatures in the
GObject namespace using a local type `value_accessor_t`.)