Define classes using common functor
The signature CLASS already describes the interface that is common to
all classes and is extended to include the function `toDerived` which
is no longer in the generated signature.

The functor Class is introduced to create a structure that implements
CLASS and code generation updated to use this functor in the generated
class structures.