Improve debug diagnostic messages
The format of messages has been made more consistent to allow machine
processing.  The module GiraffeDebug defines the types of message and
their contents.

Diagnostic messages are produced for memory operations on all kinds of
data.  Previously messages were produced only for reference counting of
objects.  The environment variable that controls output is changed from
GIRAFFFE_DEBUG_REF_COUNT to GIRAFFE_DEBUG_MEM.  The messages are logged
inside the functors Class, BoxedRecord and BoxedValueRecord which avoids
the need to generate code for logging.  These functors require the
GObject type so take an extra argument that returns the GObject type and
codegen is updated to provide this extra argument.  In the case of a
type not registered with GObject, a static string is returned containing
the C type.

A closure diagnostic message is added to log signal disconnection.  To
provide the detailed signal and associated closure in the message, the
internal representation of a handler id now includes the closure and
detailed signal.

As a result of these changes, no debug diagnostic messages are logged
from C code.