Fix handling of finalizable value in interface conversion functions
The conversion function from a class to an interface is achieved by
converting via the underlying C pointer.  Prior to this change, this
conversion followed the form for calling a foreign function, where the
foreign function was simply `I`, the identity function, which roughly
had the form:

    local
      val call =
        <Class>.FFI.withPtr false ---> <Iface>.FFI.fromPtr false
    in
      fun as<Iface> self =
        call I (<Class>.toBase self)
    end

However, when following the form for calling a foreign function, any
finalizable function argument that may be returned by the function must
be touched after the whole call to prevent premature finalization before
the `fromPtr` function has had a chance to take a reference.  As the
identity function simply returned the pointer unchanged, the argument
should have been touched.  Therefore, the following form should have
been used:

    local
      ...
    in
      fun as<Iface> self =
        call I (<Class>.toBase self) before <Class>.FFI.touchPtr self
    end

This premature finalization was observed in an application built with
MLton, causing it to crash.

As an interface conversion has a single argument, the `fromPtr` function
can be evaluated within `withPtr` and an explicit `touchPtr` is not
needed.  The fixed interface conversion function has the form:

    let
      val conv =
        <Class>.FFI.withPtr false (<Iface>.FFI.fromPtr false)
    in
      fun as<Iface> self =
        conv (<Class>.toBase self)
    end